optimize code & fix some bug #53

+++ YCDownloadSession/YCDownloadItem.h
@@ -42,8 +42,8 @@ typedef NS_ENUM(NSUInteger, YCDownloadStatus) {
 @property (nonatomic, copy, readonly, nullable) NSString *fileId;
 @property (nonatomic, copy, readonly, nonnull) NSString *downloadURL;
 @property (nonatomic, copy, readonly, nonnull) NSString *version;
-@property (nonatomic, assign, readonly) NSUInteger fileSize;
-@property (nonatomic, assign, readonly) NSUInteger downloadedSize;
+@property (nonatomic, assign, readonly) int64_t fileSize;
+@property (nonatomic, assign, readonly) int64_t downloadedSize;
 @property (nonatomic, weak, nullable) id <YCDownloadItemDelegate> delegate;
 @property (nonatomic, assign) BOOL enableSpeed;
 @property (nonatomic, strong, nullable) NSData *extraData;
